# internal/domain/user/requests.go
package user

type CreateNewUserRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
	Email    string `json:"email" binding:"required"`
}

type LoginRequest struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type SingleIdRequest struct {
	ID string `form:"id" binding:"required,uuid"`
}